+++ HISTORY.md
@@ -1,6 +1,15 @@
+v2.1.8
+===================
+* Fix demo for IE
+
 v2.1.7
 ===================
-* add es5 custom elements shim to demo page
+px-app-asset changes:
+* Add `_routeIsDifferent()` function to activatable behavior
+* Add activate tests, copied 1-for-1 from selected tests
+* Port changed code from selected behavior back to activatable behavior to fix
+  race condition where asset graph was ready earlier than expected
+
 
 v2.1.6
 ===================